Change OffsetGenerator to DateGenerator, update all files to reflect change
