+++ lib/enigma.rb
@@ -1,5 +1,5 @@
 require './lib/key_generator'
-require './lib/offset_generator'
+require './lib/date_generator'
 require './lib/caesar_cipher'
 require './lib/assignable'
 
@@ -21,7 +21,9 @@ class Enigma
   end
 
   def default_date
-    OffsetGenerator.new.date
+    DateGenerator.new.date
   end
+
+
 
 end
