+++ test/enigma_test.rb
@@ -1,6 +1,6 @@
 require './test/test_helper'
 require './lib/key_generator'
-require './lib/offset_generator'
+require './lib/date_generator'
 require './lib/caesar_cipher'
 require './lib/enigma'
