+++ test/test_helper.rb
@@ -6,7 +6,7 @@ require 'minitest/autorun'
 require 'minitest/pride'
 
 require './lib/key_generator'
-require './lib/offset_generator'
+require './lib/date_generator'
 require './lib/caesar_cipher'
 require './lib/assignable'
 require './lib/enigma'
